+++ src/components/calendar/YearView.tsx
@@ -47,7 +47,7 @@ const MiniMonth = memo(function MiniMonth({
   });
 
   return (
-    <button
+    <button type="button"
       onClick={onClick}
       data-scroll-target={containsToday || undefined}
       className={[